+++ src/containers/MyAccountMain.tsx
@@ -21,7 +21,7 @@
 
 import React from 'react';
 import { B2bSideMenu } from '../components/src';
-import RouteWithSubRoutes from '../RouteWithSubRoutes';
+import RouteWithSubRoutes from './RouteContainers/RouteWithSubRoutes';
 import { adminFetch, cortexFetch } from '../utils/Cortex';
 import * as Config from '../ep.config.json';
